Version 2.3.3
VDR developer version 2.3.3 is now available at

       ftp://ftp.tvdr.de/vdr/Developer/vdr-2.3.3.tar.bz2

A 'diff' against the previous version is available at

       ftp://ftp.tvdr.de/vdr/Developer/vdr-2.3.2-2.3.3.diff

MD5 checksums:

73182b570bcf5a67ab56f7734e479631  vdr-2.3.3.tar.bz2
112c2057dbd7e86c31f8227f61cfd2a6  vdr-2.3.2-2.3.3.diff

WARNING:
========

This is a *developer* version. Even though *I* use it in my productive
environment, I strongly recommend that you only use it under controlled
conditions and for testing and debugging.

From the HISTORY file:
- Added 'S3W ABS-3A' to sources.conf (thanks to Frank Richter).
- Fixed a possible deadlock in the recordings handler thread.
- Updated the Russian OSD texts (thanks to Andrey Pridvorov).
- Added a missing dependency to the Makefile to avoid error messages in the
  clean-plugins target (thanks to Tobias Grimm).
- The channel/CAM relations (i.e. the information which CAM can decrypt a given
  channel) are now stored in the file 'cam.data' in the cache directory (suggested
  by Dietmar Spingler). This speeds up switching to encrypted channels after
  newly starting VDR, in case there is more than one CAM in the system.
- Fixed a flaw in handling timeouts for encrypted channels.
- The mechanism of trying different CAMs when switching to an encrypted channel is
  now only triggered if there acually is more than one CAM in the system.
- Fixed updating the elapsed/remaining time in the progress display during fast
  forward/rewind.
- Changed 'unsigned' to 'signed' in some places to avoid trouble with abs() in
  gcc6+ (reported by Derek Kelly).
- CAMs that can handle multiple devices at the same time can now indicate this
  by creating the first cCamSlot as usual, and every other cCamSlot by giving
  it the first one as its "MasterSlot". To VDR this means that when searching
  for a CAM that can decrypt a particular channel, it only needs to ask the
  master CAM slot whether it is suitable for decrypting, and can skip all the
  other slots belonging to the same master. This can greatly speed up channel
  switching on systems with more than one CAM (that can handle multiple devices).
- The LCARS skin now displays the master CAM's number when a device is tuned to
  an encrypted channel.
- The Setup/CAM menu now only displays master CAMs.
- Fixed setting the local machine's SVDRP host name (was overwritten if setup.conf
  contained an empty string).
- PIDs can now be added to and deleted from a cReceiver while it is attached to
  a cDevice, without having to detach it first and re-attach it afterwards.
- Implemented support for MTD ("Multi Transponder Decryption"). This allows a CAM
  that is capable of decrypting more than one channel ("Multi Channel Decryption")
  to decrypt channels from different transponders. See the remarks in mtd.h on
  what a derived cCamSlot class needs to do in order to activate MTD (thanks to
  Jasmin Jessich for writing the ddci2 plugin and for valuable input and help
  with testing and debugging).
- The function cRingBufferLinear::Clear() can now be called safely from the
  reading thread, without additional locking.
- Now stopping any ongoing recordings before stopping the plugins, to avoid
  a crash when stopping VDR while recording.

+++ dvbdevice.c
@@ -4,7 +4,7 @@
  * See the main source file 'vdr.c' for copyright information and
  * how to reach the author.
  *
- * $Id: dvbdevice.c 4.3 2016/11/07 13:55:58 kls Exp $
+ * $Id: dvbdevice.c 4.4 2017/01/09 15:11:39 kls Exp $
  */
 
 #include "dvbdevice.h"
@@ -329,7 +329,7 @@ private:
   void ClearEventQueue(void) const;
   bool GetFrontendStatus(fe_status_t &Status) const;
   cPositioner *GetPositioner(void);
-  void ExecuteDiseqc(const cDiseqc *Diseqc, unsigned int *Frequency);
+  void ExecuteDiseqc(const cDiseqc *Diseqc, int *Frequency);
   void ResetToneAndVoltage(void);
   bool SetFrontend(void);
   virtual void Action(void);
@@ -696,7 +696,7 @@ cPositioner *cDvbTuner::GetPositioner(void)
   return positioner;
 }
 
-void cDvbTuner::ExecuteDiseqc(const cDiseqc *Diseqc, unsigned int *Frequency)
+void cDvbTuner::ExecuteDiseqc(const cDiseqc *Diseqc, int *Frequency)
 {
   if (!lnbPowerTurnedOn) {
      CHECK(ioctl(fd_frontend, FE_SET_VOLTAGE, SEC_VOLTAGE_13)); // must explicitly turn on LNB power
@@ -806,7 +806,7 @@ bool cDvbTuner::SetFrontend(void)
 
   SETCMD(DTV_DELIVERY_SYSTEM, frontendType);
   if (frontendType == SYS_DVBS || frontendType == SYS_DVBS2) {
-     unsigned int frequency = channel.Frequency();
+     int frequency = channel.Frequency();
      if (Setup.DiSEqC) {
         if (const cDiseqc *diseqc = Diseqcs.Get(device->CardIndex() + 1, channel.Source(), frequency, dtp.Polarization(), &scr)) {
            frequency -= diseqc->Lof();
@@ -829,7 +829,7 @@ bool cDvbTuner::SetFrontend(void)
         }
      else {
         int tone = SEC_TONE_OFF;
-        if (frequency < (unsigned int)Setup.LnbSLOF) {
+        if (frequency < Setup.LnbSLOF) {
            frequency -= Setup.LnbFrequLo;
            tone = SEC_TONE_OFF;
            }
